data-analyst-agent/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
pandas>=2.2.0
Pillow>=10.2.0
pyarrow>=15.0.0
numpy>=1.26.0
matplotlib>=3.8.0
